Final Year Project (Secure Online Sharing)/requirements.txt
azure-core==1.30.0
azure-storage-blob==12.19.0
bcrypt==4.1.2
blinker==1.7.0
boto3==1.34.40
botocore==1.34.40
cachelib==0.12.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
Flask==3.0.2
Flask-Session==0.6.0
google-api-core==2.17.0
google-auth==2.27.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
lxml==5.1.0
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
numpy==1.26.4
pillow==10.2.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
pyotp==2.9.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-docx==1.1.0
qrcode==7.4.2
requests==2.31.0
rsa==4.9
s3transfer==0.10.0
secure-smtplib==0.1.1
six==1.16.0
typing_extensions==4.9.0
urllib3==2.0.7
Werkzeug==3.0.1
